+++ src/lib/account-update.ts
@@ -41,7 +41,11 @@ import {
   protocolVersions,
 } from '../bindings/crypto/constants.js';
 import { MlArray } from './ml/base.js';
-import { Signature, signFieldElement } from '../mina-signer/src/signature.js';
+import {
+  Signature,
+  signFieldElement,
+  zkAppBodyPrefix,
+} from '../mina-signer/src/signature.js';
 import { MlFieldConstArray } from './ml/fields.js';
 import {
   accountUpdatesToCallForest,
@@ -65,6 +69,7 @@ import {
 } from './mina/smart-contract-context.js';
 import { assert } from './util/assert.js';
 import { RandomId } from './provable-types/auxiliary.js';
+import { NetworkId } from '../mina-signer/src/types.js';
 
 // external API
 export {
@@ -1018,9 +1023,7 @@ class AccountUpdate implements Types.AccountUpdate {
     if (Provable.inCheckedComputation()) {
       let input = Types.AccountUpdate.toInput(this);
       return hashWithPrefix(
-        activeInstance.getNetworkId() === 'mainnet'
-          ? prefixes.zkappBodyMainnet
-          : prefixes.zkappBodyTestnet,
+        zkAppBodyPrefix(activeInstance.getNetworkId()),
         packToFields(input)
       );
     } else {
@@ -1028,7 +1031,7 @@ class AccountUpdate implements Types.AccountUpdate {
       return Field(
         Test.hashFromJson.accountUpdate(
           JSON.stringify(json),
-          activeInstance.getNetworkId()
+          NetworkId.toString(activeInstance.getNetworkId())
         )
       );
     }
@@ -1399,9 +1402,7 @@ class AccountUpdate implements Types.AccountUpdate {
 function hashAccountUpdate(update: AccountUpdate) {
   return genericHash(
     AccountUpdate,
-    activeInstance.getNetworkId() === 'mainnet'
-      ? prefixes.zkappBodyMainnet
-      : prefixes.zkappBodyTestnet,
+    zkAppBodyPrefix(activeInstance.getNetworkId()),
     update
   );
 }