+++ src/lib/signature.ts
@@ -4,6 +4,7 @@ import { hashWithPrefix } from './hash.js';
 import {
   deriveNonce,
   Signature as SignatureBigint,
+  signaturePrefix,
 } from '../mina-signer/src/signature.js';
 import { Bool as BoolBigint } from '../provable/field-bigint.js';
 import {
@@ -238,6 +239,9 @@ class Signature extends CircuitValue {
   static create(privKey: PrivateKey, msg: Field[]): Signature {
     const publicKey = PublicKey.fromPrivateKey(privKey).toGroup();
     const d = privKey.s;
+    // we chose an arbitrary prefix for the signature, and it happened to be 'testnet'
+    // there's no consequences in practice and the signatures can be used with any network
+    // if there needs to be a custom nonce, include it in the message itself
     const kPrime = Scalar.fromBigInt(
       deriveNonce(
         { fields: msg.map((f) => f.toBigInt()) },
@@ -249,7 +253,7 @@ class Signature extends CircuitValue {
     let { x: r, y: ry } = Group.generator.scale(kPrime);
     const k = ry.toBits()[0].toBoolean() ? kPrime.neg() : kPrime;
     let h = hashWithPrefix(
-      prefixes.signatureTestnet,
+      signaturePrefix('testnet'),
       msg.concat([publicKey.x, publicKey.y, r])
     );
     // TODO: Scalar.fromBits interprets the input as a "shifted scalar"
@@ -265,8 +269,11 @@ class Signature extends CircuitValue {
    */
   verify(publicKey: PublicKey, msg: Field[]): Bool {
     const point = publicKey.toGroup();
+    // we chose an arbitrary prefix for the signature, and it happened to be 'testnet'
+    // there's no consequences in practice and the signatures can be used with any network
+    // if there needs to be a custom nonce, include it in the message itself
     let h = hashWithPrefix(
-      prefixes.signatureTestnet,
+      signaturePrefix('testnet'),
       msg.concat([point.x, point.y, this.r])
     );
     // TODO: Scalar.fromBits interprets the input as a "shifted scalar"